+++ examples/quickstart.js
@@ -28,8 +28,8 @@ const firstEntityValue = (entities, entity) => {
 };
 
 const actions = {
-  say: (sessionId, msg, cb) => {
-    console.log(msg);
+  say: (sessionId, context, message, cb) => {
+    console.log(message);
     cb();
   },
   merge: (sessionId, context, entities, message, cb) => {
@@ -40,8 +40,8 @@ const actions = {
     }
     cb(context);
   },
-  error: (sessionId, context) => {
-    console.log('Oops, I don\'t know what to do.');
+  error: (sessionId, context, error) => {
+    console.log(error.message);
   },
   'fetch-weather': (sessionId, context, cb) => {
     // Here should go the api call, e.g.: